+++ src/scrubbers/mean.rs
@@ -12,7 +12,7 @@ impl Scrubber for MeanScrubber {
         let sum = column
             .values()
             .filter_map(|&value| value)
-            .fold(Numeric::from(0.0), |acc, v| acc + v);
+            .fold(0.0, |acc, v| acc + v);
 
         // Grab number of all elements that are not None
         let count = column.values().filter_map(|&value| value).count();
